cloud-functions/pitch-deck-uploader/requirements.txt
functions-framework
google-cloud-storage
google-cloud-pubsub
flask
